+++ polyclonal/plot.py
@@ -836,9 +836,12 @@ def corr_heatmap(
     corr_df,
     corr_col,
     sample_cols,
+    *,
     group_col=None,
     corr_range=(0, 1),
     columns=3,
+    diverging_colors=None,
+    scheme=None,
 ):
     """Plot a correlation matrix as heat map from a tidy data frame of correlations.
 
@@ -858,6 +861,12 @@ def corr_heatmap(
         you will want to set to `(0, 1)` for :math:`r^2` and `(-1, 1)` for :math:`r`.
     columns : int
         Facet by `group_col` into this many columns.
+    diverging_colors : None or bool
+        If `True`, mid point of color scale is set to zero. If `None`, select `True`
+        if `corr_range` extends below 0.
+    scheme : None or str
+        Color scheme to use, see https://vega.github.io/vega/docs/schemes/.
+        If `None`, choose intelligently based on `corr_range` and `diverging_colors`.
 
     Returns
     -------
@@ -877,6 +886,11 @@ def corr_heatmap(
             raise ValueError(f"{corr_range[0]} > {corr_df[corr_col].min()=}")
         if corr_range[1] < corr_df[corr_col].max():
             raise ValueError(f"{corr_range[1]} < {corr_df[corr_col].max()=}")
+
+    if diverging_colors is None:
+        diverging_colors = corr_range[0] < 0
+    if scheme is None:
+        scheme = "redblue" if diverging_colors else "yellowgreenblue"
 
     if (group_col is not None) and group_col not in corr_df:
         raise ValueError(f"{group_col=} not in {corr_df.columns=}")
@@ -904,7 +918,14 @@ def corr_heatmap(
         .encode(
             x=alt.X("_label_1", title=None),
             y=alt.Y("_label_2", title=None),
-            color=alt.Color(corr_col, scale=alt.Scale(domain=corr_range)),
+            color=alt.Color(
+                corr_col,
+                scale=(
+                    alt.Scale(domainMid=0, domain=corr_range, scheme=scheme)
+                    if diverging_colors
+                    else alt.Scale(domain=corr_range, scheme=scheme)
+                ),
+            ),
             tooltip=[
                 alt.Tooltip(c, format=".3f") if corr_df[c].dtype == float else c
                 for c in corr_df.columns